modify: enhance cart UI and checkout UI
