+++ src/features/cart/components/summary/CartSummary.tsx
@@ -8,21 +8,13 @@ interface PaymentSummaryProps {
   onRemoveAll: () => void;
 }
 
-type PaymentMethod = 'stripe' | 'payos'
-
 export const PaymentSummary: React.FC<PaymentSummaryProps> = ({
   total,
   currency = '£',
   onCheckout,
   onRemoveAll,
 }) => {
-  const [selectedPayment, setSelectedPayment] = useState<PaymentMethod>('stripe');
   const [discountCode, setDiscountCode] = useState('');
-
-  const paymentMethods = [
-    { id: 'stripe' as PaymentMethod, name: 'Stripe', logo: '/images/logo/stripe.jpg' },
-    { id: 'payos' as PaymentMethod, name: 'PayOs', logo: '/images/logo/payos.png' },
-  ];
 
   const productTotal = total * 0.95; // 95% of total
   const productTax = total * 0.05; // 5% tax
@@ -42,7 +34,7 @@ export const PaymentSummary: React.FC<PaymentSummaryProps> = ({
         </div>
 
         {/* Breakdown */}
-        <div className="space-y-2 text-sm border-t pt-4">
+        <div className="space-y-2 text-sm border-t pt-4 mb-4">
           <div className="flex justify-between items-center">
             <span className="font-semibold text-gray-900">Product Total:</span>
             <span className="font-bold text-gray-900">
@@ -68,30 +60,18 @@ export const PaymentSummary: React.FC<PaymentSummaryProps> = ({
             </span>
           </div>
         </div>
-      </div>
-
-      {/* Payment Methods */}
-      <div className="mb-6">
-        <h3 className="text-gray-900 font-semibold text-lg mb-4">Payment Method</h3>
-        <div className="grid grid-cols-2 gap-3">
-          {paymentMethods.map((method) => (
-            <button
-              key={method.id}
-              onClick={() => setSelectedPayment(method.id)}
-              className={`p-4 rounded-lg border-2 transition-all hover:scale-105 ${
-                selectedPayment === method.id
-                  ? 'border-blue-500 bg-blue-50 shadow-md'
-                  : 'border-gray-200 bg-white hover:border-gray-300'
-              }`}
-            >
-              <div className="flex flex-col items-center gap-2">
-                <Image src={method.logo} alt='payment-icon' width={64} height={64} />
-                {/* <span className="text-sm font-medium text-gray-700">
-                  {method.name}
-                </span> */}
-              </div>
-            </button>
-          ))}
+        <div className='space-y-2 text-sm border-t pt-4'>
+          <p className="text-sm mb-8">
+          By placing your order you agree to the website
+          <br/>
+          <a href="#" className="underline hover:text-blue-500 transition-colors hover:text-semibold">
+            Terms & Conditions
+          </a>
+          {' '}&{' '}
+          <a href="#" className="underline hover:text-blue-500 transition-colors hover:text-semibold">
+            Privacy notice
+          </a>
+        </p>
         </div>
       </div>
